backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.24.3
pydantic==2.5.0
python-jwt==4.0.0
bcrypt==4.1.2
python-multipart==0.0.6
websockets==12.0
Pillow==10.1.0